+++ src/main/java/org/akip/service/dto/TaskInstanceDTO.java
@@ -1,6 +1,5 @@
 package org.akip.service.dto;
 
-import org.akip.camunda.form.CamundaFormFieldDef;
 import org.akip.domain.enumeration.StatusTaskInstance;
 import org.akip.domain.enumeration.TypeTaskInstance;
 
@@ -57,9 +56,9 @@ public class TaskInstanceDTO implements Serializable {
 
     private String connectorConfigName;
 
-    private List<CamundaFormFieldDef> formFields;
-
     private ProcessDefinitionDTO processDefinition;
+
+    private TaskDefinitionDTO taskDefinition;
 
     private ProcessInstanceDTO processInstance;
 
@@ -223,14 +222,6 @@ public class TaskInstanceDTO implements Serializable {
         this.connectorConfigName = connectorConfigName;
     }
 
-    public List<CamundaFormFieldDef> getFormFields() {
-        return formFields;
-    }
-
-    public void setFormFields(List<CamundaFormFieldDef> formFields) {
-        this.formFields = formFields;
-    }
-
     public ProcessDefinitionDTO getProcessDefinition() {
         return processDefinition;
     }
@@ -247,35 +238,35 @@ public class TaskInstanceDTO implements Serializable {
         this.processInstance = processInstance;
     }
 
+    public TaskDefinitionDTO getTaskDefinition() {
+        return taskDefinition;
+    }
+
+    public void setTaskDefinition(TaskDefinitionDTO taskDefinition) {
+        this.taskDefinition = taskDefinition;
+    }
+
     @Override
     public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof TaskInstanceDTO)) {
-            return false;
-        }
-
-        TaskInstanceDTO taskInstanceDTO = (TaskInstanceDTO) o;
-        if (this.id == null) {
-            return false;
-        }
-        return Objects.equals(this.id, taskInstanceDTO.id);
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        TaskInstanceDTO that = (TaskInstanceDTO) o;
+        return Objects.equals(id, that.id);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(this.id);
+        return Objects.hash(id);
     }
 
     // prettier-ignore
+
     @Override
     public String toString() {
         return "TaskInstanceDTO{" +
-            "id=" + getId() +
-            ", taskId='" + getTaskId() + "'" +
-            ", name='" + getName() + "'" +
-            ", status='" + getStatus() + "'" +
-            "}";
+                "id=" + id +
+                ", taskId='" + taskId + '\'' +
+                ", name='" + name + '\'' +
+                '}';
     }
 }